refactor(mir): split monolithic ir_generation into modular lowering architecture
- Extract 2638-line ir_generation.rs into focused modules under lowering/
- Implement trait-based architecture with LowerExpr and LowerStmt
- Integrate optimization passes (FuseCmpBranch, InPlaceOptimization, etc.)
- Fix duplicate function bug in MIR generation
- Preserve all existing optimizations and functionality
- All 51 tests passing
